// dream/src/app/models/ApplicationDetails.ts
export class ApplicationDetails {

	firstname: string;
	lastname: string;
	middlename:string;
	applicationId: number;
	email: string;
	phoneNo: string;
	dateOfBirth: string;
	nationality: string;
	aadharNo: string;
	panNo: string;
	dateOfAppointment: string;
	applicationStatusMessage: string;

	accountNo: number;
	ammount:number;
	accountType:string;

	status: boolean;

}