// dream/src/app/models/LoanFields.ts
export class LoanFields {
    loanId: number;
    interestRate: number;
    loanAmount: number;
    maxLoanAmount: number;
    tenure: number;
    emi: number;
    customerMonthlyIncome: number;
    eligibilityStatus: number;
    startDate: string;
    endDate: string;
    loanStatus: string;
    applicationId: number;
    statusMessage: string;
    status: boolean;
}